Refactored POJO packages
