+++ app/src/main/java/org/db_poultry/db/flockDAO/ReadFlock.java
@@ -1,8 +1,8 @@
 package org.db_poultry.db.flockDAO;
 
-import org.db_poultry.pojo.Flock;
-import org.db_poultry.pojo.FlockComplete;
-import org.db_poultry.pojo.FlockDetails;
+import org.db_poultry.pojo.FlockPOJO.Flock;
+import org.db_poultry.pojo.FlockPOJO.FlockComplete;
+import org.db_poultry.pojo.FlockPOJO.FlockDetails;
 
 import java.sql.*;
 import java.util.ArrayList;
