+++ app/src/main/java/org/db_poultry/pojo/FlockPOJO/FlockComplete.java
@@ -1,4 +1,4 @@
-package org.db_poultry.pojo;
+package org.db_poultry.pojo.FlockPOJO;
 
 import java.util.ArrayList;
 import java.util.List;
