+++ app/src/main/java/org/db_poultry/pojo/SupplyPOJO/SupplyComplete.java
@@ -1,4 +1,4 @@
-package org.db_poultry.pojo;
+package org.db_poultry.pojo.SupplyPOJO;
 
 import java.sql.Date;
